src: separate nibble abstraction in a dedicated header file
