+++ src/cpu/reorder/simple_reorder.hpp
@@ -24,6 +24,7 @@
 #include "common/c_types_map.hpp"
 #include "common/dnnl_thread.hpp"
 #include "common/math_utils.hpp"
+#include "common/nibble.hpp"
 #include "common/primitive.hpp"
 #include "common/primitive_attr.hpp"
 #include "common/tag_traits.hpp"
